Restructure project files. Updated NuGet packages Added Frody and ErrorProne.NET.CoreAnalyzers
