--- AtosArrows.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AtosArrows", "AtosArrows.csproj", "{8D64CBD3-0566-4149-8C9C-FFB8BAA863B0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8D64CBD3-0566-4149-8C9C-FFB8BAA863B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8D64CBD3-0566-4149-8C9C-FFB8BAA863B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8D64CBD3-0566-4149-8C9C-FFB8BAA863B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8D64CBD3-0566-4149-8C9C-FFB8BAA863B0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {205232EB-96BF-418C-8D3E-33234BF45774}
-	EndGlobalSection
-EndGlobal
