--- packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="HarmonyX" version="2.4.2" targetFramework="net462" />
-  <package id="JotunnLib" version="2.0.12" targetFramework="net462" />
-  <package id="Mono.Cecil" version="0.11.3" targetFramework="net462" />
-  <package id="MonoMod" version="21.4.21.3" targetFramework="net462" />
-  <package id="MonoMod.RuntimeDetour" version="21.4.21.3" targetFramework="net462" />
-  <package id="MonoMod.Utils" version="21.4.21.3" targetFramework="net462" />
-  <package id="Pfhoenix.Valheim.ModProjectReferences" version="1.0.6" targetFramework="net462" />
-</packages>
