+++ src/AtosArrowsJVL/FodyWeavers.xsd
@@ -1,0 +1,101 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
+  <!-- This file was generated by Fody. Manual changes to this file will be lost when your project is rebuilt. -->
+  <xs:element name="Weavers">
+    <xs:complexType>
+      <xs:all>
+        <xs:element name="ILMerge" minOccurs="0" maxOccurs="1">
+          <xs:complexType>
+            <xs:all>
+              <xs:element name="IncludeAssemblies" type="xs:string" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A regular expression matching the assembly names to include in merging.</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+              <xs:element name="ExcludeAssemblies" type="xs:string" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A regular expression matching the assembly names to exclude from merging.</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+              <xs:element name="IncludeResources" type="xs:string" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A regular expression matching the resource names to include in merging.</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+              <xs:element name="ExcludeResources" type="xs:string" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A regular expression matching the resource names to exclude from merging.</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+              <xs:element name="HideImportedTypes" type="xs:boolean" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A switch to control whether the imported types are hidden (made private) or keep their visibility unchanged. Default is 'true'</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+              <xs:element name="NamespacePrefix" type="xs:string" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A string that is used as prefix for the namespace of the imported types.</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+              <xs:element name="FullImport" type="xs:boolean" minOccurs="0" maxOccurs="1">
+                <xs:annotation>
+                  <xs:documentation>A switch to control whether to import the full assemblies or only the referenced types. Default is 'false'</xs:documentation>
+                </xs:annotation>
+              </xs:element>
+            </xs:all>
+            <xs:attribute name="IncludeAssemblies" type="xs:string">
+              <xs:annotation>
+                <xs:documentation>A regular expression matching the assembly names to include in merging.</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="ExcludeAssemblies" type="xs:string">
+              <xs:annotation>
+                <xs:documentation>A regular expression matching the assembly names to exclude from merging.</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="IncludeResources" type="xs:string">
+              <xs:annotation>
+                <xs:documentation>A regular expression matching the resource names to include in merging.</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="ExcludeResources" type="xs:string">
+              <xs:annotation>
+                <xs:documentation>A regular expression matching the resource names to exclude from merging.</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="HideImportedTypes" type="xs:boolean">
+              <xs:annotation>
+                <xs:documentation>A switch to control whether the imported types are hidden (made private) or keep their visibility unchanged. Default is 'true'</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="NamespacePrefix" type="xs:string">
+              <xs:annotation>
+                <xs:documentation>A string that is used as prefix for the namespace of the imported types.</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="FullImport" type="xs:boolean">
+              <xs:annotation>
+                <xs:documentation>A switch to control whether to import the full assemblies or only the referenced types. Default is 'false'</xs:documentation>
+              </xs:annotation>
+            </xs:attribute>
+          </xs:complexType>
+        </xs:element>
+      </xs:all>
+      <xs:attribute name="VerifyAssembly" type="xs:boolean">
+        <xs:annotation>
+          <xs:documentation>'true' to run assembly verification (PEVerify) on the target assembly after all weavers have been executed.</xs:documentation>
+        </xs:annotation>
+      </xs:attribute>
+      <xs:attribute name="VerifyIgnoreCodes" type="xs:string">
+        <xs:annotation>
+          <xs:documentation>A comma-separated list of error codes that can be safely ignored in assembly verification.</xs:documentation>
+        </xs:annotation>
+      </xs:attribute>
+      <xs:attribute name="GenerateXsd" type="xs:boolean">
+        <xs:annotation>
+          <xs:documentation>'false' to turn off automatic generation of the XML Schema file.</xs:documentation>
+        </xs:annotation>
+      </xs:attribute>
+    </xs:complexType>
+  </xs:element>
+</xs:schema>
