+++ src/AtosArrowsJVL/packages.config
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ErrorProne.NET.CoreAnalyzers" version="0.1.2" targetFramework="net462" developmentDependency="true" />
+  <package id="Fody" version="6.5.2" targetFramework="net462" developmentDependency="true" />
+  <package id="HarmonyX" version="2.5.1" targetFramework="net462" />
+  <package id="ILMerge.Fody" version="1.16.0" targetFramework="net462" />
+  <package id="JotunnLib" version="2.1.3" targetFramework="net462" />
+  <package id="Mono.Cecil" version="0.11.4" targetFramework="net462" />
+  <package id="MonoMod" version="21.7.8.3" targetFramework="net462" />
+  <package id="MonoMod.RuntimeDetour" version="21.7.8.3" targetFramework="net462" />
+  <package id="MonoMod.Utils" version="21.7.8.3" targetFramework="net462" />
+  <package id="Pfhoenix.Valheim.ModProjectReferences" version="1.0.6" targetFramework="net462" />
+</packages>
